polars_benchmark/requirements.txt
polars==0.20.10
pyarrow==15.0.0
pandas==2.2.0
gcsfs==2023.12.2
google-cloud-storage==2.14.0
